Split out player routes to separate folder
